requirements.txt
python-telegram-bot==13.5
python-dotenv==0.17.0
dialogflow==1.1.0.
google-cloud-dialogflow==2.2.0
vk-api==11.9.4